Added current work so far.
Added all current work. It needs a bit more to be completed in order to run a debug version in Visual Studio. Also, placeholder spider_dynamic_01 and Servo h and cpp file are included. The placeholders need to be removed and the ino file needs to be created with the code in the spider_dynamic_01.cpp file.